% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.analysis.R
\name{factor.analysis}
\alias{factor.analysis}
\title{Factor Analysis}
\usage{
factor.analysis(X, factors, categories = NULL, rotation = "none")
}
\arguments{
\item{X}{A dataframe of quantitative data}

\item{factors}{The number of factors to use in the factor analysis. Must be
greater than 1 but less than the number of variables in the dataset.}

\item{categories}{An optional dataframe with one column representing a
categorial variable on the observations in X. Used for coloring the scatter
plot. Can be set to NULL to disable this.}

\item{rotation}{Rotation specifier. See documentation for `factanal`.}
}
\value{
The output of the constructed `factanal` call.
}
\description{
A wrapper for the `factanal` function that provides input checking, and plots
of the results.
}
\examples{
fa <- factor.analysis(freedom_index[4:15], factors = 3,
                      categories = freedom_index[2], rotation = "promax")
}
