% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.count.R
\name{factor.count}
\alias{factor.count}
\title{Factor Count Analysis}
\usage{
factor.count(X = NULL, S = NULL, scale = FALSE)
}
\arguments{
\item{X}{A dataframe of quantitative data. Optional if S is
specified.}

\item{S}{An optional variance-covariance matrix that supersedes X if not set
to NULL.}

\item{scale}{Whether to use standardized variables in the analysis.}
}
\value{
A list containing the variances for each component and a cumulative
sum of these variances. Also contains the automatically found elbow in the
scree plot and the angle at each point. Finally, reports the number of
components with variance exceeding unit variance, and those exceeding 0.9
and 1.1 variance, respectively.
}
\description{
Plots a scree plot and cumulative sum of variance plot, and calculates
quantities helpful to choosing the number of components and/or factor to
use in a PCA or factor analysis.
}
\examples{
factor_info <- factor.count(X = freedom_index[4:15])
}
